// client/src/models/bois/BoisRouge.ts
export interface PostBoisRouge {
    type: string;
    marque: string;
    fornisseur: string;
    n_fardou: string;
    epaisseur: number;
    larg: number;
    pieces: number;
    metre_lineare: number;
    prix_unity: number;
    volume: number;
    long_moyenne: number;
    quality: string;
    date_creation: string;
    long: number;
    piece: number;
  }
  export interface GetBoisRouge {
    id: number;
    type: string;
    marque: string;
    fornisseur: string;
    n_fardou: string;
    epaisseur: number;
    larg: number;
    pieces: number;
    metre_lineare: number;
    prix_unity: number;
    volume: number;
    long_moyenne: number;
    quality: string;
    date_creation: string;
    long: number;
    piece: number;
  }
 